homework 5.14 Modulation
