+++ index.html
@@ -10,8 +10,7 @@
 </head>
 
 <body>
-  <script src="/js/functions.js"></script>
-  <script src = "/js/main.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -233,6 +232,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
+
+  <script src = "/js/main.js" type = "module"></script>
 
 </body>
 </html>
